chore: allow to specify several endpoints
This change relies on protobuf spec, where it doesn't matter how many times the field appears in the byte stream, because decoder will always take the last one for non-repeated fields. We use it to our advantage and simply pass primary endpoint address at the end, but we take it as the first.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>